// eco-assist/src/types/types.ts
export type TipoCliente = {
  nome_cliente: string;
  cpf_cliente: string;
  geladeira: string;
};

export type TipoEmpresa = {
  nome_empresa: string;
  cnpj_empresa: string;
  computador_ou_notebook: string;
  podencia_computador_notebook: string;
  tempo_uso_computador_notebook: string;
  ar_condicionado: string;
  podencia_ar_condicionado: string;
  tempo_uso_ar_condicionado: string;
  motores_eletrico: string;
  podencia_motores_eletrico: string;
  tempo_uso_motores_eletrico: string;
  sistema_ventilacao: string;
  podencia_sistema_ventilacao: string;
  tempo_uso_sistema_ventilacao: string;
  sistema_refrigeracao: string;
  podencia_sistema_refrigeracao: string;
  tempo_uso_sistema_refrigeracao: string;
  equipamento_embalagem: string;
  podencia_equipamento_embalagem: string;
  tempo_uso_equipamento_embalagem: string;
};